requirements.txt
beautifulsoup4==4.7.1
bs4==0.0.1
certifi==2019.3.9
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.6.0
diff-match-patch==20181111
django-cors-headers==3.2.0
django-crispy-forms==1.8.1
django-filter==2.2.0
django-formtools==2.2
django-guardian==2.1.0
django-import-export==2.0
django-redis==4.10.0
django-reversion==3.0.5
django-tinymce==2.4.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
drf-extensions==0.5.0
et-xmlfile==1.0.1
future==0.18.2
html5lib==1.0.1
httplib2==0.9.2
idna==2.8
itchat==1.3.10
itypes==1.1.0
jdcal==1.4
Jinja2==2.10.3
lxml==4.3.3
Markdown==3.1.1
MarkupPy==1.14
MarkupSafe==1.1.1
mysqlclient==1.4.6
Naked==0.1.31
oauthlib==3.1.0
odfpy==1.4.0
openpyxl==2.5.12
pbr==5.1.1
Pillow==5.4.1
pycryptodome==3.9.4
pygame==1.9.6
PyJWT==1.7.1
PyMySQL==0.9.3
pypng==0.0.19
PyQRCode==1.2.1
python3-openid==3.1.0
pytz==2018.9
PyYAML==5.2
redis==3.3.11
requests==2.21.0
requests-oauthlib==1.3.0
selenium==3.141.0
shellescape==3.4.1
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
soupsieve==1.9
sqlparse==0.3.0
stevedore==1.30.0
tablib==0.14.0
uiautomator==0.3.6
uritemplate==3.0.1
urllib3==1.24.1
virtualenv==16.2.0
virtualenv-clone==0.4.0
virtualenvwrapper==4.8.2
virtualenvwrapper-win==1.2.5
webencodings==0.5.1
workbook==1.1
xlrd==1.2.0
XlsxWriter==1.2.6
xlutils==2.0.0
xlwt==1.3.0
